Cast uid_t to (long) to print it, silencing warnings on Solaris

+++ http.c
@@ -445,13 +445,13 @@ static int run_csd_script(struct openconnect_info *vpninfo, char *buf, int bufle
 			struct passwd *pw;
 
 			if (setuid(vpninfo->uid_csd)) {
-				fprintf(stderr, _("Failed to set uid %d\n"),
-					vpninfo->uid_csd);
+				fprintf(stderr, _("Failed to set uid %ld\n"),
+					(long)vpninfo->uid_csd);
 				exit(1);
 			}
 			if (!(pw = getpwuid(vpninfo->uid_csd))) {
-				fprintf(stderr, _("Invalid user uid=%d\n"),
-					vpninfo->uid_csd);
+				fprintf(stderr, _("Invalid user uid=%ld\n"),
+					(long)vpninfo->uid_csd);
 				exit(1);
 			}
 			setenv("HOME", pw->pw_dir, 1);
